Update cl_luabee.lua
UPDATED:
function LUABEE.GetValidArgs( ... )
I'm not sure of your purpose for GetValidArgs, I'm assuming you want it to go through arguments until it reaches a nil, or false and then cut off ANY REMAINING AS WELL. I recreated that with an iteration which is cleaner, shorter, and doesn't have a limit of input arguments.


function string.Safe( text )
I updated the string.Safe to be shorter, cleaner, while encompassing ALL potentially unsafe characters. I included the URL in the comments. %w matches all alphanumeric so a-zA-Z0-9, when you capitalize any in that list, it does the INVERSE, so it matches anything that is NOT the aforementioned pattern and replaces them with _. You could theoretically return the gsub without editing the variable to remove O( 3 ) which reduces to O( 1 ) but because string.gsub returns text, howManyReplacementsWereMadeAsANumber if you were to print( ), you would have a number hanging off on the side. I'm unsure of how you use this, so I didn't want to change your return. 

Hope this helps,
-Josh